Fix dtype of some Inplace params of optimizer_lamb_phase1 (fix for bf16 tests)
The dtype of input params wt_norm, adam_norm, and adam_step were not
inferred from params type when created. This led to the output type of
some syn_tensors to be misinterpreted at kernel building.

This fix uses the proper params type for the tensors mentioned and that
fixes the bf16 SingleOp tests

+++ python_packages/habana_frameworks/torch/hpex/optimizers/FusedLamb.py
@@ -173,9 +173,9 @@ class FusedLamb(Optimizer):
                 exp_avg_list.append(exp_avg)
                 exp_avg_sq_list.append(exp_avg_sq)
 
-                wt_norm_list.append(torch.empty((1,), device=self.device))
-                adam_norm_list.append(torch.empty((1,), device=self.device))
-                adam_step_list.append(torch.empty_like(exp_avg))
+                wt_norm_list.append(torch.empty((1,), device=self.device).to(p.dtype))
+                adam_norm_list.append(torch.empty((1,), device=self.device).to(p.dtype))
+                adam_step_list.append(torch.empty_like(exp_avg).to(p.dtype))
 
             torch.ops.hpu.optimizer_lamb_phase1(
                 grad_list,
